% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_matrix}
\alias{aggregate_matrix}
\title{Aggregate matrix over groups}
\usage{
aggregate_matrix(x, groups = NULL, fun = Matrix::colMeans)
}
\arguments{
\item{groups}{A character vector with the groups to aggregate over.}

\item{fun}{The aggregation function to be applied to each chunk of the matrix.}
}
\value{
A summary matrix.
}
\description{
Aggregate matrix over groups
}
