% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{annotation_plot}
\alias{annotation_plot}
\title{Plot brain structure annotation}
\usage{
annotation_plot(annot_df, annotation_level = "custom_2",
  annotation_colors = many, show_coordinates = F, show_legend = F,
  alpha = 0.5)
}
\description{
Plot brain structure annotation
}
