% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_map.R
\name{assign_cells}
\alias{assign_cells}
\alias{assign_cells.VoxelMap}
\title{Assign cells to structure}
\usage{
assign_cells(object, ...)

\method{assign_cells}{VoxelMap}(object)
}
\arguments{
\item{object}{A VoxelMap opject}
}
\value{
A tibble with cells assigned to voxels
}
\description{
Assign cells to structure
}
