% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{brain_annotation_plot}
\alias{brain_annotation_plot}
\title{Plot brain annotation}
\usage{
brain_annotation_plot(stage = "E13pt5", annotation_level = "custom_2",
  annotation_colors = many, show_coordinates = F, show_legend = F,
  alpha = 0.5)
}
\description{
Plot brain annotation
}
