% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{brain_expression_plot}
\alias{brain_expression_plot}
\title{Plot gene expression across the mouse brain}
\usage{
brain_expression_plot(expr_mat, stage = "E13pt5", genes = NULL)
}
\description{
Plot gene expression across the mouse brain
}
