% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{brain_plot}
\alias{brain_plot}
\title{Plot correlation to entire brain from saggital view}
\usage{
brain_plot(corr_df, corr_colors = rdpu, newpage = T)
}
\description{
Plot correlation to entire brain from saggital view
}
