% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainspan_map.R, R/generics.R
\name{brainspan_map.default}
\alias{brainspan_map.default}
\alias{brainspan_map.Seurat}
\alias{brainspan_map}
\title{Project single cell gene expression data to BrainSpan}
\usage{
\method{brainspan_map}{default}(
  object,
  stages = c(8:16),
  groups = NULL,
  method = "pearson",
  genes_use = NULL,
  allow_neg = FALSE,
  pseudobulk_groups = FALSE
)

\method{brainspan_map}{Seurat}(
  object,
  stages = c(8:16),
  group_name = NULL,
  method = "pearson",
  genes_use = NULL,
  allow_neg = FALSE,
  pseudobulk_groups = FALSE
)

brainspan_map(object, ...)
}
\arguments{
\item{groups}{A character or factor vector or for grouping of cells,
e.g. clusters, cell types.}

\item{method}{A character string indicating which correlation coefficient to compute.}

\item{genes_use}{A character vector with genes to use for computing the correlation.
We recommend to use 150 - 500 genes.}

\item{allow_neg}{Logical. Whether to allow negative correlations or set them to 0.}

\item{pseudobulk_groups}{Logical. Whether to summarizse the group expression before computing the correlation.}

\item{group_name}{A string indicating the metadata column for grouping the cells,
e.g. clusters, cell types.}

\item{stage}{A integer vector indicating the stages (pcw) to map to..}
}
\value{
A BrainSpanMap object with a cell x ref correlation matrix and metadata.
}
\description{
Calculates correlation between single-cell gene expression and reference samples from
BrainSpan
}
