% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{de}
\alias{de}
\title{Performs differential expression analysis on a sample x feature matrix}
\usage{
de(expr_mat, groups)
}
\value{
A tibble with p-values and AUROC values.
}
\description{
Performs differential expression analysis on a sample x feature matrix
}
