% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R, R/generics.R
\name{deconvolute.default}
\alias{deconvolute.default}
\alias{deconvolute}
\title{Devonvolute bulk RNA-seq data based on reference data}
\usage{
\method{deconvolute}{default}(
  expr_mat,
  genes_use,
  reference = "ABA",
  stage = "E13",
  annotation_level = "custom_2",
  pseudo_tpm = FALSE,
  other_types = FALSE,
  involve_regions = NULL
)

deconvolute(object, ...)
}
\arguments{
\item{expr_mat}{Expression matrix to deconvolute.}

\item{genes_use}{Marker genes to use for deconvolution.}

\item{reference}{Which reference to use.}

\item{stage}{Developmental stage to use as reference.}

\item{other_types}{Whether to infer proportions of type 'other'.}
}
\value{
A tibble with deconvoluted proportions for each reference structure.
}
\description{
Devonvolute bulk RNA-seq data based on reference data
}
