% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_aba_data}
\alias{load_aba_data}
\title{Define the path to voxel data and load lazily if required}
\usage{
load_aba_data(dir, lazy = T)
}
\description{
Define the path to voxel data and load lazily if required
}
