% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_mousebrain_data}
\alias{load_mousebrain_data}
\title{Load the mousebrain data (LaManno & Siletti et al. 2020)}
\usage{
load_mousebrain_data(dir)
}
\arguments{
\item{dir}{The directory with loom files for each stage.}
}
\description{
Load the mousebrain data (LaManno & Siletti et al. 2020)
}
