% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_message}
\alias{log_message}
\title{Print diagnostic message}
\usage{
log_message(text, verbose = TRUE)
}
\arguments{
\item{text}{Text to print}

\item{verbose}{Whether to print}
}
\description{
Print diagnostic message
}
