% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mousebrain_map.R
\name{mousebrain_map}
\alias{mousebrain_map}
\alias{mousebrain_map.default}
\alias{mousebrain_map.Seurat}
\title{Project single cell gene expression data to developing mouse brain data (LaManno & Siletti et al. 2020)}
\usage{
mousebrain_map(object, ...)

\method{mousebrain_map}{default}(
  object,
  groups = NULL,
  method = "pearson",
  genes_use = NULL,
  allow_neg = FALSE,
  pseudobulk_groups = FALSE
)

\method{mousebrain_map}{Seurat}(
  object,
  group_name = NULL,
  method = "pearson",
  genes_use = NULL,
  allow_neg = FALSE,
  pseudobulk_groups = FALSE
)
}
\arguments{
\item{groups}{A character or factor vector or for grouping of cells,
e.g. clusters, cell types.}

\item{method}{A character string indicating which correlation coefficient to compute.}

\item{genes_use}{A character vector with genes to use for computing the correlation.
We recommend to use 150 - 500 genes.}

\item{allow_neg}{Logical. Whether to allow negative correlations or set them to 0.}

\item{pseudobulk_groups}{Logical. Whether to summarizse the group expression before computing the correlation.}

\item{group_name}{A string indicating the metadata column for grouping the cells,
e.g. clusters, cell types.}
}
\value{
A MousebrainMap object with a cell x ref correlation matrix and metadata.
}
\description{
Calculates correlation between single-cell gene expression and clusters from LaManno & Siletti et al. 2020
}
