% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_annotation}
\alias{plot_annotation}
\alias{plot_annotation.default}
\alias{plot_annotation.VoxelMap}
\title{Plot brain structure annotation}
\usage{
plot_annotation(object = "E13", ...)

\method{plot_annotation}{default}(object = "E13",
  annotation_level = "custom_2", annotation_colors = many,
  slices = NULL, show_coordinates = F, show_legend = F,
  alpha = 0.5)

\method{plot_annotation}{VoxelMap}(object, annotation_level = "custom_2",
  annotation_colors = many, slices = NULL, show_coordinates = F,
  show_legend = F, alpha = 0.5)
}
\description{
Plot brain structure annotation

Plot brain structure annotation

Plot brain structure annotation
}
