% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_expression}
\alias{plot_expression}
\title{Plot gene expression across the mouse brain}
\usage{
plot_expression(stage = "E13", genes = NULL)
}
\description{
Plot gene expression across the mouse brain
}
