% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_expression_3d}
\alias{plot_expression_3d}
\title{Plot gene expression across the mouse brain in 3D}
\usage{
plot_expression_3d(stage, gene, annotation_level = NULL,
  annotation_colors = many, expression_colors = inferno,
  sizes = c(10, 1000), both_hemispheres = T, ...)
}
\description{
Plot gene expression across the mouse brain in 3D
}
