% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{project_to_brain}
\alias{project_to_brain}
\title{Map correlation matrix to brain grid and plot the result}
\usage{
project_to_brain(corr_mat, cell_groups = NULL)
}
\description{
Map correlation matrix to brain grid and plot the result
}
