% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{project_to_slices}
\alias{project_to_slices}
\title{Map correlation matrix to brain slices and plot the result}
\usage{
project_to_slices(corr_mat, cell_groups = NULL, slices = c(6, 11, 23,
  28, 36), annotation_level = "level_1", annotation_colors = many)
}
\description{
Map correlation matrix to brain slices and plot the result
}
