% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_loom}
\alias{read_loom}
\title{Read loom files to list}
\usage{
read_loom(filename, layer = "matrix", row_id = "voxel",
  col_id = "gene")
}
\arguments{
\item{filename}{The name of the loom file}

\item{layer}{The name of the layer to load as matrix}

\item{row_id}{A name for the colum in the row_meta indicating the row names.}

\item{col_id}{A name for the colum in the col_meta indicating the column names.}
}
\description{
Read loom files to list
}
