% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/reference_map.R
\name{reference_map}
\alias{reference_map}
\alias{reference_map.default}
\alias{reference_map.Seurat}
\title{Compare a query scRNA-seq dataset with a reference}
\usage{
reference_map(object, ...)

\method{reference_map}{default}(
  object,
  reference_object,
  query_meta = NULL,
  reference_meta = NULL,
  groups = NULL,
  reduction = NULL,
  method = "pearson",
  genes_use = NULL,
  allow_neg = FALSE,
  pseudobulk_groups = TRUE
)

\method{reference_map}{Seurat}(
  object,
  reference_object,
  slot = "data",
  assay = "RNA",
  group_name = NULL,
  reduction = "umap",
  method = "pearson",
  genes_use = NULL,
  allow_neg = FALSE,
  pseudobulk_groups = TRUE
)
}
\arguments{
\item{object}{A matrix with query expression data.}

\item{reference_object}{A matrix with reference expression data.}

\item{reference_meta}{A data frame with reference metadata.}

\item{groups}{A character or factor vector or for grouping of cells,
e.g. clusters, cell types.}

\item{method}{A character string indicating which correlation coefficient to compute.}

\item{genes_use}{A character vector with genes to use for computing the correlation.
We recommend to use 150 - 500 genes.}

\item{allow_neg}{Logical. Whether to allow negative correlations or set them to 0.}

\item{pseudobulk_groups}{Logical. Whether to summarize the group expression before computing the correlation.}

\item{group_name}{A string indicating the metadata column for grouping the cells,
e.g. clusters, cell types.}
}
\value{
A ReferenceMap object with a cell x cell correlation matrix and metadata.
}
\description{
Calculates correlation between single-cell gene expression.
}
