% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{run_epic}
\alias{run_epic}
\title{Run EPIC deconvolution}
\usage{
run_epic(expr_mat, ref_expr, genes_use, ref_sd = NULL, other_types = FALSE)
}
\arguments{
\item{expr_mat}{Input matrix to deconvolute.}

\item{ref_expr}{Reference expression matrix.}

\item{genes_use}{Marker genes to use.}

\item{ref_sd}{Variability of reference expression.}

\item{other_types}{Whether to infer proportions of type 'other'.}
}
\value{
A tibble with deconvoluted proportions for each reference structure.
}
\description{
Run EPIC deconvolution
}
