% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{similarity_bar_plot}
\alias{similarity_bar_plot}
\title{Plot correlation to individual brain structures as bar plot}
\usage{
similarity_bar_plot(similarity_df, annotation_level = "custom_3",
  annotation_colors = many)
}
\arguments{
\item{annotation_level}{The structure annotation level to color code.}

\item{annotation_colors}{Color map for structure annotation.}
}
\description{
Plot correlation to individual brain structures as bar plot
}
