% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{slice}
\alias{slice}
\alias{slice.VoxelMap}
\title{Slice VoxelMap and plot result}
\usage{
slice(object, ...)

\method{slice}{VoxelMap}(object, slices = c(6, 11, 23, 28, 36),
  groups = NULL, annotation_level = "level_1",
  annotation_colors = many)
}
\description{
Slice VoxelMap and plot result

Slice VoxelMap and plot result
}
