% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparse_cor}
\alias{sparse_cor}
\title{Fast correlation for sparse matrices}
\usage{
sparse_cor(x, y = NULL)

sparse_cor(x, y = NULL)
}
\description{
Fast correlation for sparse matrices

Fast correlation for sparse matrices
}
