% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparse_covcor}
\alias{sparse_covcor}
\title{Fast correlation and covariance calcualtion for sparse matrices}
\usage{
sparse_covcor(x, y = NULL)
}
\description{
Fast correlation and covariance calcualtion for sparse matrices
}
