% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stage_name}
\alias{stage_name}
\title{Pick proper stage name}
\usage{
stage_name(stage)
}
\arguments{
\item{stage}{An alias for the developmental stage.}
}
\value{
A valid stage name.
}
\description{
Pick proper stage name
}
