% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{structure_markers}
\alias{structure_markers}
\title{Finds differentially expressed features for all structures on a given annotation level}
\usage{
structure_markers(expr_mat, annotation_level = NULL,
  structure_groups = NULL)
}
\description{
Finds differentially expressed features for all structures on a given annotation level
}
