% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_map.R
\name{summarise_groups}
\alias{summarise_groups}
\alias{summarise_groups.VoxelMap}
\title{Summarize VoxelMap object over groups}
\usage{
summarise_groups(object, ...)

\method{summarise_groups}{VoxelMap}(object, groups, fun = colMeans)
}
\description{
Summarize VoxelMap object over groups

Summarize VoxelMap object over groups
}
