% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_map.R
\name{summarise_structures}
\alias{summarise_structures}
\alias{summarise_structures.VoxelMap}
\title{Summarize VoxelMap object over structures}
\usage{
summarise_structures(object, ...)

\method{summarise_structures}{VoxelMap}(object,
  annotation_level = "custom_3", fun = Matrix::colMeans)
}
\description{
Summarize VoxelMap object over structures

Summarize VoxelMap object over groups
}
