% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainspan_map.R, R/generics.R,
%   R/mousebrain_map.R, R/voxel_map.R
\name{summarize_structures.BrainSpanMap}
\alias{summarize_structures.BrainSpanMap}
\alias{summarize_structures}
\alias{summarize_structures.MousebrainMap}
\alias{summarize_structures.VoxelMap}
\title{Summarize data over structures}
\usage{
\method{summarize_structures}{BrainSpanMap}(
  object,
  annotation_level = c("structure_group", "structure_name", "structure_acronym"),
  fun = colMeans
)

summarize_structures(object, ...)

\method{summarize_structures}{MousebrainMap}(
  object,
  annotation_level = c("region", "class"),
  fun = colMeans
)

\method{summarize_structures}{VoxelMap}(object, annotation_level = "custom_3", fun = colMeans)
}
\arguments{
\item{annotation_level}{The structure annotation level to summarize voxels to.}

\item{fun}{Function to use for summarizing voxels.}
}
\value{
A tibble with structure summaries

A tibble with structure summaries

A tibble with structure summaries
}
\description{
Summarize data over structures
}
