% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{three_dim_plot}
\alias{three_dim_plot}
\title{3D intensity plot}
\usage{
three_dim_plot(int_df, annotation_level = NULL,
  annotation_colors = many, intensity_colors = inferno)
}
\description{
3D intensity plot
}
