% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{variable_genes}
\alias{variable_genes}
\title{Finds differentially expressed features for all structures on a given annotation level}
\usage{
variable_genes(stage = "E13pt5", method = "vst", nfeatures = 2000)
}
\description{
Finds differentially expressed features for all structures on a given annotation level
}
