% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{voxel_correlation}
\alias{voxel_correlation}
\alias{voxel_correlation.default}
\alias{voxel_correlation.Seurat}
\title{Correlate single-cell gene expression with in situ hybridization data}
\usage{
voxel_correlation(object, ...)

\method{voxel_correlation}{default}(object, stage = "E13",
  groups = NULL, method = "pearson", allow_neg = F,
  genes_use = NULL)

\method{voxel_correlation}{Seurat}(object, stage = "E13",
  group_name = NULL, method = "pearson", allow_neg = F,
  genes_use = NULL)
}
\description{
Correlate single-cell gene expression with in situ hybridization data

Correlate single-cell gene expression with in situ hybridization data

Correlate single-cell gene expression with in situ hybridization data
}
