% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zscale}
\alias{zscale}
\title{Scale function that works for vectors and grouped dataframes}
\usage{
zscale(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A scales numeric vector.
}
\description{
Scale function that works for vectors and grouped dataframes
}
