% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{annotation_plot}
\alias{annotation_plot}
\title{Plot brain structure annotation}
\usage{
annotation_plot(annot_df, annotation_level = "custom_2",
  annotation_colors = many, show_coordinates = F, show_legend = F,
  alpha = 0.5)
}
\arguments{
\item{annotation_level}{The structure annotation level to color code.}

\item{annotation_colors}{A character vector with the color scale.}

\item{show_coordinates}{Logical. Whether to show slice coordinates or not.}

\item{show_legend}{Logical. Whether to show a color legend or not.}

\item{alpha}{Float. The alpha value for grid tiles.}
}
\value{
An annotation plot.
}
\description{
Plot brain structure annotation
}
