% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_map.R
\name{assign_cells}
\alias{assign_cells}
\alias{assign_cells.VoxelMap}
\title{Assign cells to structure}
\usage{
assign_cells(object, ...)

\method{assign_cells}{VoxelMap}(object)
}
\description{
Assign cells to structure

Assign cells to structure based on maximum correlation voxel
}
