% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/reference_map.R
\name{brainspan_map}
\alias{brainspan_map}
\alias{brainspan_map.default}
\alias{brainspan_map.Seurat}
\title{Project single cell gene expression data to BrainSpan}
\usage{
brainspan_map(object, ...)

\method{brainspan_map}{default}(object, stages = c(8:16),
  groups = NULL, method = "pearson", genes_use = NULL,
  allow_neg = F)

\method{brainspan_map}{Seurat}(object, stages = c(8:16),
  group_name = NULL, method = "pearson", genes_use = NULL,
  allow_neg = F)
}
\arguments{
\item{groups}{A character or factor vector or for grouping of cells,
e.g. clusters, cell types.}

\item{method}{A character string indicating which correlation coefficient to compute.}

\item{genes_use}{A character vector with genes to use for computing the correlation.
We recommend to use 150 - 500 genes.}

\item{allow_neg}{Logical. Whether to allow negative correlations or set them to 0.}

\item{group_name}{A string indicating the metadata column for grouping the cells,
e.g. clusters, cell types.}

\item{stage}{A integer vector indicating the stages (pcw) to map to..}
}
\value{
A ReferenceMap object with a cell x ref correlation matrix and metadata.
}
\description{
Calculates correlation between single-cell gene expression reference samples from
BrainSpan
}
