% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{feature_plot}
\alias{feature_plot}
\title{Feature plot}
\usage{
feature_plot(expr_mat, meta, markers, plot = xy_plot, title = NULL,
  ncol = NULL, nrow = NULL, sort = T, scale = T)
}
\description{
Feature plot
}
