% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_markers}
\alias{get_markers}
\title{Get gene expression values from matrix}
\usage{
get_markers(expr_mat, markers, scale = TRUE)
}
\arguments{
\item{expr_mat}{A samples x genes gene expression matrix.}

\item{markers}{A character vector with the names of the genes to obtain.}

\item{scale}{Logical. Whether to scale the expression values.}
}
\value{
A tibble with expression of selected genes.
}
\description{
Get gene expression values from matrix
}
