% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_markers}
\alias{get_markers}
\title{Get gene expression from matrix and return data frame}
\usage{
get_markers(expr_mat, markers, scale = T)
}
\description{
Get gene expression from matrix and return data frame
}
