% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mapping_plot}
\alias{mapping_plot}
\title{Plot correlation to entire brain from saggital view}
\usage{
mapping_plot(map_df, view = "saggital", map_colors = rdpu,
  newpage = T)
}
\description{
Plot correlation to entire brain from saggital view
}
