% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mapping_plot}
\alias{mapping_plot}
\title{Plot correlation to entire brain or selected slices.}
\usage{
mapping_plot(map_df, view = "sagittal", slices = NULL,
  map_colors = rdpu, ...)
}
\arguments{
\item{view}{String indicating the perspective to show. Valid values are
'sagittal', 'coronal', 'traverse', 'z' , 'x', 'y', 'slice', '3D'.}

\item{slices}{A numeric vector indicating the slices to plot.}

\item{map_colors}{Color map for correlation values.}

\item{...}{Other arguments passed to egg::ggarrange().}
}
\value{
A similarity map.
}
\description{
Plot correlation to entire brain or selected slices.
}
