% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_annotation}
\alias{plot_annotation}
\alias{plot_annotation.default}
\alias{plot_annotation.VoxelMap}
\title{Plot brain structure annotation}
\usage{
plot_annotation(object = "E13", ...)

\method{plot_annotation}{default}(object = "E13",
  annotation_level = "custom_2", annotation_colors = many,
  slices = NULL, show_coordinates = F, show_legend = F,
  alpha = 0.5)

\method{plot_annotation}{VoxelMap}(object, annotation_level = "custom_2",
  annotation_colors = many, slices = NULL, show_coordinates = F,
  show_legend = F, alpha = 0.5)
}
\arguments{
\item{object}{String indicating the developmental stage from the ABA.}

\item{annotation_level}{The structure annotation level to color code.}

\item{annotation_colors}{A character vector with the color scale}

\item{slices}{A numeric vector indicating the slices to plot.}

\item{show_coordinates}{Logical. Whether to show slice coordinates or not.}

\item{show_legend}{Logical. Whether to show a color legend or not.}

\item{alpha}{Float. The alpha value for grid tiles.}
}
\description{
Plot brain structure annotation

Plot brain structure annotation

Plot brain structure annotation
}
