% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plots.R
\name{plot_map}
\alias{plot_map}
\alias{plot_map.VoxelMap}
\alias{plot_map.BrainSpanMap}
\alias{plot_map.MousebrainMap}
\title{Plot similarity map of single cells}
\usage{
plot_map(object, ...)

\method{plot_map}{VoxelMap}(
  object,
  view = c("sagittal", "coronal", "traverse", "z", "x", "y", "slice", "3D"),
  slices = NULL,
  groups = NULL,
  annotation_level = "custom_2",
  annotation_colors = many,
  map_colors = gyrdpu_flat,
  show_coordinates = FALSE,
  show_legend = FALSE,
  ...
)

\method{plot_map}{BrainSpanMap}(object, groups = NULL, map_colors = blues, scale = FALSE)

\method{plot_map}{MousebrainMap}(
  object,
  groups = NULL,
  map_colors = ylorrd_flat,
  point_size = 0.2,
  subsample = 50000,
  scale = T,
  show_legend = FALSE,
  ...
)
}
\arguments{
\item{...}{Other arguments passed to patchwork::wrap_plots().}

\item{view}{String indicating the perspective to show. Valid values are
'sagittal', 'coronal', 'traverse', 'z' , 'x', 'y', 'slice', '3D'.}

\item{slices}{A numeric vector indicating the slices to plot.}

\item{groups}{A metadata column or character vector to group the cells,
e.g. clusters, cell types.}

\item{annotation_level}{The structure annotation level to summarize to.}

\item{annotation_colors}{Color map for structure annotation.}

\item{map_colors}{Color map for correlation values.}

\item{show_coordinates}{Logical. Whether to show slice coordinates or not.}

\item{show_legend}{Logical. Whether to show a color legend or not.}

\item{scale}{Logical. Whether to scale correlation values.}

\item{point_size}{Point size.}

\item{subsample}{Subsample mousebrain cells for faster plotting.}
}
\value{
A similarity map.

A similarity map.

A similarity map.
}
\description{
Plot similarity map of single cells
}
