% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_map_3d}
\alias{plot_map_3d}
\alias{plot_map_3d.VoxelMap}
\title{Plot similarity map of single cells to voxels in 3D}
\usage{
plot_map_3d(object, ...)

\method{plot_map_3d}{VoxelMap}(object, groups = NULL,
  show_group = NULL, annotation_level = NULL,
  annotation_colors = many, map_colors = inferno_flat, sizes = c(1,
  1000), both_hemispheres = T, ...)
}
\description{
Plot similarity map of single cells to voxels in 3D

Plot similarity map of single cells to voxels in 3D
}
