% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plots.R
\name{plot_map_3d}
\alias{plot_map_3d}
\alias{plot_map_3d.VoxelMap}
\title{Plot similarity map of single cells to voxels in 3D}
\usage{
plot_map_3d(object, ...)

\method{plot_map_3d}{VoxelMap}(
  object,
  groups = NULL,
  show_group = NULL,
  annotation_level = NULL,
  annotation_colors = NULL,
  map_colors = inferno_flat,
  sizes = c(1, 1000),
  both_hemispheres = TRUE,
  ...
)
}
\arguments{
\item{...}{Other arguments passed to plotly::plot_ly().}

\item{groups}{A metadata column or character vector to group the cells,
e.g. clusters, cell types.}

\item{annotation_level}{The structure annotation level to color code.}

\item{annotation_colors}{Color map for structure annotation.}

\item{map_colors}{Color map for correlation values.}

\item{sizes}{The size range for points.}

\item{both_hemispheres}{Logical. Whether to plot both hemispheres (TRUE) of only one (FALSE).}

\item{group_show}{The nameof the group to show in the plot.}
}
\value{
A similarity map in 3D.
}
\description{
Plot similarity map of single cells to voxels in 3D
}
