% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_structure_similarity}
\alias{plot_structure_similarity}
\alias{plot_structure_similarity.VoxelMap}
\title{Plot similarity to brain structures}
\usage{
plot_structure_similarity(object, ...)

\method{plot_structure_similarity}{VoxelMap}(object, groups = NULL,
  annotation_level = "custom_3", annotation_groups = "custom_2",
  annotation_colors = many, type = "box")
}
\value{
A plot showing average correlation to each brain structure.
}
\description{
Plot similarity to brain structures
}
