% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/plots.R
\name{plot_structure_similarity}
\alias{plot_structure_similarity}
\alias{plot_structure_similarity.VoxelMap}
\alias{plot_structure_similarity.MousebrainMap}
\title{Plot similarity to brain structures}
\usage{
plot_structure_similarity(object, ...)

\method{plot_structure_similarity}{VoxelMap}(
  object,
  groups = NULL,
  annotation_level = "custom_3",
  annotation_colors = blues_flat,
  cluster = TRUE,
  include_unannotated = FALSE,
  scale = T,
  ...
)

\method{plot_structure_similarity}{MousebrainMap}(
  object,
  groups = NULL,
  annotation_level = c("class", "region"),
  region_colors = many,
  class_colors = many,
  heat_colors = blues_flat,
  cluster = TRUE,
  scale = T,
  ...
)
}
\arguments{
\item{groups}{A metadata column or character vector to group the cells,
e.g. clusters, cell types.}

\item{annotation_level}{The structure annotation level to color code.}

\item{annotation_colors}{Color map for structure annotation.}

\item{cluster}{Logical. Perform hierarchical clustering on rows and columns.}
}
\value{
A plot showing average correlation to each brain structure.

A plot showing average correlation to each brain structure.
}
\description{
Plot similarity to brain structures
}
