% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainspan_map.R, R/mousebrain_map.R,
%   R/voxel_map.R
\name{print.BrainSpanMap}
\alias{print.BrainSpanMap}
\alias{print.MousebrainMap}
\alias{print.VoxelMap}
\title{Print BrainSpanMap objects}
\usage{
\method{print}{BrainSpanMap}(object)

\method{print}{MousebrainMap}(object)

\method{print}{VoxelMap}(object)
}
\description{
Print BrainSpanMap objects

Print MousebrainMap objects

Print VoxelMap objects
}
