% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_map.R, R/voxel_map.R
\name{print.ReferenceMap}
\alias{print.ReferenceMap}
\alias{print.VoxelMap}
\title{Print ReferenceMap objects}
\usage{
\method{print}{ReferenceMap}(object)

\method{print}{VoxelMap}(object)
}
\description{
Print ReferenceMap objects

Print VoxelMap objects
}
