% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_cor}
\alias{safe_cor}
\title{Safe correlation function which returns a sparse matrix without missing values}
\usage{
safe_cor(x, y, method = "pearson", allow_neg = F)
}
\description{
Safe correlation function which returns a sparse matrix without missing values
}
