% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{slice_plot}
\alias{slice_plot}
\title{Plot correlation to brain slices and structure annotation}
\usage{
slice_plot(slice_df, annotation_level = "custom_2",
  annotation_colors = many, map_colors = gyrdpu, ...)
}
\description{
Plot correlation to brain slices and structure annotation
}
