% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{structure_markers}
\alias{structure_markers}
\title{Finds differentially expressed features for all structures on a given annotation level}
\usage{
structure_markers(stage = "E13", annotation_level = "custom_3",
  structure_groups = NULL)
}
\arguments{
\item{stage}{The developmental stage from the ABA.}

\item{annotation_level}{The structure annotation level to perform DE an analysis between.}

\item{structure_groups}{A character or factor vector to provide custom grouping of voxels.}
}
\description{
Finds differentially expressed features for all structures on a given annotation level
}
