% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_map.R
\name{summarise_groups}
\alias{summarise_groups}
\title{Correlate single-cell gene expression with in situ hybridization data}
\usage{
summarise_groups(object, groups, fun = colMeans)

\method{summarise_groups}{VoxelMap}(object, groups, fun = colMeans)
}
\description{
Correlate single-cell gene expression with in situ hybridization data

Summarize VoxelMap object over groups
}
