% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_map.R
\name{summarize_groups}
\alias{summarize_groups}
\alias{summarize_groups.VoxelMap}
\title{Summarize VoxelMap object over groups}
\usage{
summarize_groups(object, ...)

\method{summarize_groups}{VoxelMap}(object, groups = NULL,
  fun = colMeans)
}
\arguments{
\item{groups}{A metadata column or character vector to group the cells,
e.g. clusters, cell types.}

\item{fun}{Function used to aggregate the groups.}
}
\value{
A tibble with group summaries
}
\description{
Summarize VoxelMap object over groups
}
