% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainspan_map.R, R/generics.R,
%   R/mousebrain_map.R, R/voxel_map.R
\name{summarize_groups.BrainSpanMap}
\alias{summarize_groups.BrainSpanMap}
\alias{summarize_groups}
\alias{summarize_groups.MousebrainMap}
\alias{summarize_groups.VoxelMap}
\title{Summarize data over groups}
\usage{
\method{summarize_groups}{BrainSpanMap}(object, groups = NULL, fun = colMeans)

summarize_groups(object, ...)

\method{summarize_groups}{MousebrainMap}(object, groups = NULL, fun = colMeans)

\method{summarize_groups}{VoxelMap}(object, groups = NULL, fun = colMeans)
}
\arguments{
\item{groups}{A metadata column or character vector to group the cells,
e.g. clusters, cell types.}

\item{fun}{Function used to aggregate the groups.}
}
\value{
A tibble with group summaries

A tibble with group summaries

A tibble with group summaries
}
\description{
Summarize data over groups
}
