% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_map.R
\name{summarize_structures}
\alias{summarize_structures}
\alias{summarize_structures.VoxelMap}
\title{Summarize VoxelMap object over structures}
\usage{
summarize_structures(object, ...)

\method{summarize_structures}{VoxelMap}(object,
  annotation_level = "custom_3", fun = colMeans)
}
\arguments{
\item{annotation_level}{The structure annotation level to summarize voxels to.}

\item{fun}{Function to use for summarizing voxels.}
}
\description{
Summarize VoxelMap object over structures

Summarize VoxelMap object over groups
}
