% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/reference_map.R, R/voxel_map.R
\name{summarize_structures}
\alias{summarize_structures}
\alias{summarize_structures.ReferenceMap}
\alias{summarize_structures.VoxelMap}
\title{Summarize data over structures}
\usage{
summarize_structures(object, ...)

\method{summarize_structures}{ReferenceMap}(
  object,
  annotation_level = c("structure_group", "structure_name", "structure_acronym"),
  fun = colMeans
)

\method{summarize_structures}{VoxelMap}(object, annotation_level = "custom_3", fun = colMeans)
}
\arguments{
\item{annotation_level}{The structure annotation level to summarize voxels to.}

\item{fun}{Function to use for summarizing voxels.}
}
\value{
A tibble with structure summaries

A tibble with structure summaries
}
\description{
Summarize data over structures
}
