% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{three_dim_plot}
\alias{three_dim_plot}
\title{General 3D plotting function}
\usage{
three_dim_plot(
  int_df,
  annotation_level = NULL,
  annotation_colors = many,
  intensity_colors = inferno,
  sizes = c(10, 1000),
  ...
)
}
\arguments{
\item{annotation_level}{The structure annotation level to color code.}

\item{annotation_colors}{Color map for structure annotation.}

\item{sizes}{The size range for points.}

\item{...}{Other arguments passed to plotly::plot_ly().}

\item{intnesity_colors}{Colors for intensity scale.}

\item{both_hemispheres}{Logical. Whether to plot both hemispheres (TRUE) of only one (FALSE).}
}
\value{
A three-dimensional plot.
}
\description{
General 3D plotting function
}
