% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{variable_genes}
\alias{variable_genes}
\title{Finds differentially expressed features for all structures on a given annotation level}
\usage{
variable_genes(stage = "E13", nfeatures = 2000)
}
\arguments{
\item{stage}{A string indicating the developmental stage in the ABA.}

\item{nfeatures}{The number of features to return.}
}
\description{
Finds differentially expressed features for all structures on a given annotation level
}
