% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{voxel_correlation}
\alias{voxel_correlation}
\title{Correlate single-cell gene expression with in situ hybridization data}
\usage{
voxel_correlation(expr_mat, voxel_mat, allow_neg = F,
  method = "pearson", genes_use = NULL)
}
\description{
Correlate single-cell gene expression with in situ hybridization data
}
