% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/voxel_map.R
\name{voxel_map}
\alias{voxel_map}
\alias{voxel_map.default}
\alias{voxel_map.Seurat}
\title{Project single cell gene expression data to spatial expression patterns in the mouse brain}
\usage{
voxel_map(object, ...)

\method{voxel_map}{default}(
  object,
  stage = "E13",
  groups = NULL,
  method = "pearson",
  genes_use = NULL,
  allow_neg = FALSE,
  pseudobulk_groups = TRUE
)

\method{voxel_map}{Seurat}(
  object,
  stage = "E13",
  group_name = NULL,
  method = "pearson",
  genes_use = NULL,
  allow_neg = FALSE,
  pseudobulk_groups = TRUE
)
}
\arguments{
\item{stage}{The developmental stage in the ABA to map single cells to.}

\item{groups}{A character or factor vector or for grouping of cells,
e.g. clusters, cell types.}

\item{method}{A character string indicating which correlation coefficient to compute.}

\item{genes_use}{A character vector with genes to use for computing the correlation.
We recommend to use 150 - 500 genes.}

\item{allow_neg}{Logical. Whether to allow negative correlations or set them to 0.}

\item{pseudobulk_groups}{Logical. Whether to summarizse the group expression before computing the correlation.}

\item{group_name}{A string indicating the metadata column for grouping the cells,
e.g. clusters, cell types.}
}
\value{
A VoxelMap object with a cell x voxel correlation matrix and metadata.
}
\description{
Calculates correlation between single-cell gene expression and
in situ hybridization data of the developing mouse brain from
the Allen Brain Atlas.
}
