% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_map.R
\name{voxel_map}
\alias{voxel_map}
\alias{voxel_map.default}
\alias{voxel_map.Seurat}
\title{Correlate single-cell gene expression with in situ hybridization data}
\usage{
voxel_map(object, ...)

\method{voxel_map}{default}(object, stage = "E13", groups = NULL,
  method = "pearson", allow_neg = F, genes_use = NULL)

\method{voxel_map}{Seurat}(object, stage = "E13", group_name = NULL,
  method = "pearson", allow_neg = F, genes_use = NULL)
}
\description{
Correlate single-cell gene expression with in situ hybridization data

Correlate single-cell gene expression with in situ hybridization data

Correlate single-cell gene expression with in situ hybridization data
}
